Expect executive comp fields to be moved from recipient.officers to latest_transaction.contract_data
